test(image-generation): move Lite resolution regression to its own file
runtime.test.ts exceeded the 1000-line test-file lint ceiling after the
regression was appended; move it to a focused runtime-resolution-clamp
test file instead of suppressing max-lines.
